requirements.txt
Django==2.2.18
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.15.0
Pillow==8.1.0
djangorestframework==3.10.3
dj-database-url==0.5.0
drf-flex-fields==0.7.5
djoser==2.1.0
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
drf-yasg==1.20.0
django-webpack-loader==0.7.0
django-silk==4.1.0
django-extensions==3.1.1
requests==2.25.1
django-redis==4.12.1
celery==5.0.5
django-celery-results== 2.0.1
time-machine==2.0.1
django-admin-display==1.3.0
django-jazzmin==2.4.4
